Refactoring and making CreateNodesUrlUtils
