+++ frontend/src/components/Workspace/FlowChart/RightDrawer/AlgorithmParamForm.tsx
@@ -6,6 +6,7 @@ import { Tooltip } from "@mui/material"
 
 import { createParamFormItemComponent } from "components/common/ParamFormItemCreator"
 import { SectionTitle } from "components/common/ParamSection"
+import { useCreateNodesUrl } from "components/utils/CreateNodesUrlUtils"
 import { ParamFormContext } from "components/Workspace/FlowChart/RightDrawer/ParamFormContents"
 import { getAlgoParams } from "store/slice/AlgorithmNode/AlgorithmNodeActions"
 import {
@@ -29,29 +30,9 @@ export const AlgorithmParamForm = memo(function AlgorithmParamForm() {
     selectAlgorithmParamsKeyList(nodeId),
     arrayEqualityFn,
   )
-  const algoNameMapping: { [key: string]: string } = {
-    eta: "eta-event-triggered-average",
-    cca: "cca-canonical-correlation-analysis",
-    dpca: "dpca-demixed-principal-component-analysis",
-    dca: "dca-dynamical-component-analysis",
-    tsne: "tsne-t-distributed-stochastic-neighbor-embedding",
-    glm: "glm-generalized-linear-model",
-    lda: "lda-linear-discriminant-analysis",
-    svm: "svm-support-vector-machine",
-    granger: "granger-granger-causality-test",
-    "lccd-cell-detection": "lccd-detect",
-    "microscope-to-img": "microscope-to-image",
-    "cnmf-multisession": "caiman-cnmf-multisession",
-  }
 
-  let formattedAlgoName = algoName.toLowerCase().replace(/_/g, "-")
+  const parameterUrl = useCreateNodesUrl(algoName)
 
-  // Check if the formatted name exists in the mapping, otherwise keep it as is
-  formattedAlgoName =
-    algoNameMapping[formattedAlgoName as keyof typeof algoNameMapping] ||
-    formattedAlgoName
-
-  const parameterUrl = `https://optinist.readthedocs.io/en/latest/specifications/algorithm_nodes.html#${formattedAlgoName}`
   useEffect(() => {
     if (!algoParamIsLoaded) {
       dispatch(getAlgoParams({ nodeId, algoName }))
